# docs/api/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "AURA Care-Chain API",
        "description": "API REST do AURA Care-Chain \u2014 assistente de sa\u00fade domiciliar voice-first\npara idosos com Parkinson, com cadeia log\u00edstica de seguran\u00e7a da casa.\n\nConsumida por tr\u00eas clientes: app Flutter (Android/iOS/Web),\napp React Native e painel administrativo Angular.\n\nRegras de ouro: nunca prescreve nem diagnostica, o escore \u00e9 sempre\nexplic\u00e1vel (fatores + pesos) e nenhum dado de sa\u00fade \u00e9 gravado sem\nconsentimento LGPD.\n",
        "contact": {
            "name": "Equipe AURA \u2014 FIAP Smart HAS"
        },
        "license": {
            "name": "Uso acad\u00eamico \u2014 FIAP 2026"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
        "/api/v1/homes/{homeId}/checklist": {
            "put": {
                "tags": [
                    "2. Casas"
                ],
                "summary": "Atualiza o checklist de seguran\u00e7a (alimenta o escore)",
                "operationId": "checklist",
                "parameters": [
                    {
                        "name": "homeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChecklistRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChecklistResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha inesperada \u2014 a mensagem interna fica no log do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "INTERNAL_ERROR": {
                                        "value": {
                                            "error": {
                                                "code": "INTERNAL_ERROR",
                                                "message": "Erro interno.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente, inv\u00e1lido ou expirado \u2014 ou credenciais incorretas no login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "UNAUTHORIZED": {
                                        "value": {
                                            "error": {
                                                "code": "UNAUTHORIZED",
                                                "message": "Autentica\u00e7\u00e3o necess\u00e1ria.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "NOT_FOUND": {
                                        "value": {
                                            "error": {
                                                "code": "NOT_FOUND",
                                                "message": "Casa n\u00e3o encontrada(a).",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Papel sem permiss\u00e3o (RBAC) ou recurso de outro paciente (isolamento)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "FORBIDDEN": {
                                        "value": {
                                            "error": {
                                                "code": "FORBIDDEN",
                                                "message": "Acesso negado a este recurso.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/homes": {
            "get": {
                "tags": [
                    "2. Casas"
                ],
                "summary": "Lista as casas vis\u00edveis ao usu\u00e1rio autenticado",
                "operationId": "list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HomeResponse"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha inesperada \u2014 a mensagem interna fica no log do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "INTERNAL_ERROR": {
                                        "value": {
                                            "error": {
                                                "code": "INTERNAL_ERROR",
                                                "message": "Erro interno.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente, inv\u00e1lido ou expirado \u2014 ou credenciais incorretas no login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "UNAUTHORIZED": {
                                        "value": {
                                            "error": {
                                                "code": "UNAUTHORIZED",
                                                "message": "Autentica\u00e7\u00e3o necess\u00e1ria.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "NOT_FOUND": {
                                        "value": {
                                            "error": {
                                                "code": "NOT_FOUND",
                                                "message": "Casa n\u00e3o encontrada(a).",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Papel sem permiss\u00e3o (RBAC) ou recurso de outro paciente (isolamento)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "FORBIDDEN": {
                                        "value": {
                                            "error": {
                                                "code": "FORBIDDEN",
                                                "message": "Acesso negado a este recurso.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Consentimento LGPD n\u00e3o aceito (RN-001)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "CONSENT_REQUIRED": {
                                        "value": {
                                            "error": {
                                                "code": "CONSENT_REQUIRED",
                                                "message": "Aceite a Pol\u00edtica de Privacidade antes de registrar dados de sa\u00fade.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "2. Casas"
                ],
                "summary": "Cria a casa e resolve o endere\u00e7o pelo CEP (ViaCEP)",
                "operationId": "create_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateHomeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HomeCreatedResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha inesperada \u2014 a mensagem interna fica no log do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "INTERNAL_ERROR": {
                                        "value": {
                                            "error": {
                                                "code": "INTERNAL_ERROR",
                                                "message": "Erro interno.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente, inv\u00e1lido ou expirado \u2014 ou credenciais incorretas no login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "UNAUTHORIZED": {
                                        "value": {
                                            "error": {
                                                "code": "UNAUTHORIZED",
                                                "message": "Autentica\u00e7\u00e3o necess\u00e1ria.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "NOT_FOUND": {
                                        "value": {
                                            "error": {
                                                "code": "NOT_FOUND",
                                                "message": "Casa n\u00e3o encontrada(a).",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Papel sem permiss\u00e3o (RBAC) ou recurso de outro paciente (isolamento)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "FORBIDDEN": {
                                        "value": {
                                            "error": {
                                                "code": "FORBIDDEN",
                                                "message": "Acesso negado a este recurso.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Consentimento LGPD n\u00e3o aceito (RN-001)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "CONSENT_REQUIRED": {
                                        "value": {
                                            "error": {
                                                "code": "CONSENT_REQUIRED",
                                                "message": "Aceite a Pol\u00edtica de Privacidade antes de registrar dados de sa\u00fade.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ChecklistRequest": {
                "required": [
                    "items"
                ],
                "type": "object",
                "properties": {
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object",
                            "example": {
                                "grab_bar_bathroom": false,
                                "slippery_floor": true
                            }
                        },
                        "example": {
                            "grab_bar_bathroom": false,
                            "slippery_floor": true
                        }
                    }
                }
            },
            "ChecklistResponse": {
                "type": "object",
                "properties": {
                    "safetyChecklist": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object"
                        }
                    }
                }
            },
...
            "CreateHomeRequest": {
                "required": [
                    "patientName"
                ],
                "type": "object",
                "properties": {
                    "patientName": {
                        "type": "string"
                    },
                    "birthDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "cep": {
                        "pattern": "\\d{8}",
                        "type": "string",
                        "example": "01310100"
                    },
                    "label": {
                        "type": "string"
                    }
                }
            },
            "HomeCreatedResponse": {
                "type": "object",
                "properties": {
                    "homeId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "address": {
                        "type": "string"
                    },
                    "lat": {
                        "type": "number",
                        "format": "double"
                    },
                    "lng": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
...
            "HomeResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "label": {
                        "type": "string"
                    },
                    "patientName": {
                        "type": "string"
                    },
                    "birthDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "cep": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "lat": {
                        "type": "number",
                        "format": "double"
                    },
                    "lng": {
                        "type": "number",
                        "format": "double"
                    },
                    "safetyChecklist": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object"
                        }
                    }
                }
            },
...
            "ApiErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/Body"
                    }
                }
            },
            "Body": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "details": {
                        "type": "object"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido em POST /api/v1/auth/login",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}